+++ main.py
@@ -477,10 +477,10 @@ def main() -> None:
             sys.argv = start + shlex.split(cfg.BOT.LAUNCH_OPTIONS) + end
             App = BotApp
         case "craft":
-            sys.argv = start + shlex.split(cfg.MOVE.LAUNCH_OPTIONS) + end
+            sys.argv = start + shlex.split(cfg.CRAFT.LAUNCH_OPTIONS) + end
             App = CraftApp
         case "move":
-            sys.argv = start + shlex.split(cfg.CRAFT.LAUNCH_OPTIONS) + end
+            sys.argv = start + shlex.split(cfg.MOVE.LAUNCH_OPTIONS) + end
             App = MoveApp
         case "harvest":
             sys.argv = start + shlex.split(cfg.HARVEST.LAUNCH_OPTIONS) + end