+++ docs/en/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed FishHookedError freezing bug
 - Fixed friction brake process bug
 - Fixed bottom fishing recast bug
+- Fixed hookset bug
 
 ## 0.7.1 (2025-08-16)