+++ docs/en/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix harvest bug
 - Add hookset
+- Improve default config
 
 ### 🐛 Bug Fixes
 - Fixed a critical bug that caused the script to panic when a fish was caught during the resetting phase.