+++ docs/en/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix harvest bug
 - Add hookset
 - Improve default config
+- Improve float detection and float configs
 
 ### 🐛 Bug Fixes
 - Fixed a critical bug that caused the script to panic when a fish was caught during the resetting phase.