+++ docs/en/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improve default config
 - Improve float detection and float configs
 - Fix telescopic mode reset bug
+- Fix spin mode bait check bug
 
 ### 🐛 Bug Fixes
 - Fixed a critical bug that caused the script to panic when a fish was caught during the resetting phase.